Moved lots of game models to the game manager sessions module
